♻️(services) refactor Services into a single directory (#304)
* 🐛(scalingo) add an env var to disable Celery beat scheduling

Deployments on fresh PaaS databases would fail because of this, without
giving the user the ability to create the database through a manual
run command.

* 🐛(nginx) fix error when no env vars configured

* ♻️(services) refactor Services into a single directory

I haven't done AI yet because there are pending patches but once done we'll also have core/services/ai/*.py
I used importer as a name to avoid using a Python keyword

* 🚨(lint) move imports to the top
