+++ src/backend/core/api/viewsets/send.py
@@ -15,7 +15,7 @@ from rest_framework.views import APIView
 
 from core import models
 from core.mda.outbound import prepare_outbound_message
-from core.tasks import send_message_task
+from core.mda.tasks import send_message_task
 
 from .. import permissions, serializers
