+++ src/backend/core/services/importer/__init__.py
@@ -1,0 +1,5 @@
+"""Import service for handling message imports."""
+
+from .service import ImportService
+
+__all__ = ["ImportService"]
