+++ src/backend/core/signals.py
@@ -8,12 +8,12 @@ from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 
 from core import models
-from core.identity.keycloak import (
+from core.services.identity.keycloak import (
     sync_mailbox_to_keycloak_user,
     sync_maildomain_to_keycloak_group,
 )
-from core.search import MESSAGE_INDEX, get_opensearch_client
-from core.tasks import index_message_task, reindex_thread_task
+from core.services.search import MESSAGE_INDEX, get_opensearch_client
+from core.services.search.tasks import index_message_task, reindex_thread_task
 
 logger = logging.getLogger(__name__)
