+++ src/nginx/servers.conf.erb
@@ -26,7 +26,7 @@ server {
 	}
     
     # Django admin
-    location ^~ /<%= ENV["DJANGO_ADMIN_URL"] %> {
+    location ^~ /<%= ENV["DJANGO_ADMIN_URL"] || "admin/" %> {
         proxy_set_header X-Forwarded-Proto https;
         proxy_set_header Host $http_host;
         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
